its-arbin-time: switched check-package-lock over to npm arborist (instead of manually parsing)
